Minor fix to Reciever and added tests for Warlord
