+++ src/main/kotlin/game/collections/Army.kt
@@ -54,4 +54,7 @@ class Army : Receiver() {
         units[position].equipWeapon(weapon)
     }
 
+
 }
+
+
